% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representation_conversions.R
\name{as.polm}
\alias{as.polm}
\alias{as.polm.lpolm}
\title{Coerece to polynomial object}
\usage{
as.polm(obj, ...)

\method{as.polm}{lpolm}(obj, ...)
}
\arguments{
\item{obj}{\code{\link{lpolm}} object}

\item{...}{other arguments}
}
\value{
\code{\link{polm}} object
}
\description{
Coerece to polynomial object
}
\examples{
lp = lpolm(1:3, min_deg = 1) 
as.polm(lp)
}
