% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_methods.R
\name{as.stsp}
\alias{as.stsp}
\alias{as.stsp.lpolm}
\alias{as.stsp.polm}
\alias{as.stsp.lmfd}
\alias{as.stsp.rmfd}
\alias{as.stsp.pseries}
\title{Coerce to Statespace Realization}
\usage{
as.stsp(obj, ...)

\method{as.stsp}{lpolm}(obj, ...)

\method{as.stsp}{polm}(obj, ...)

\method{as.stsp}{lmfd}(obj, ...)

\method{as.stsp}{rmfd}(obj, ...)

\method{as.stsp}{pseries}(obj, method = c("balanced", "echelon"), ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional additional parameters}

\item{method}{character string}
}
\value{
object of class \code{\link{stsp}}.
}
\description{
The function \code{\link{as.stsp.pseries}} calls
\code{\link{pseries2stsp}} with default parameters.
Of course the \code{\link{pseries}} object must contain
sufficiently many lags.
}
\examples{
(rr = test_rmfd(dim = c(3,2), degrees = c(2,1)))
as.stsp(rr)
}
