% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_print.R
\name{as_tex_matrixfilter}
\alias{as_tex_matrixfilter}
\title{TeX Matrix Polynomial Filters}
\usage{
as_tex_matrixfilter(coefs, x = "z", t = "t")
}
\arguments{
\item{coefs}{3-dimensional array with the coefficients of the filter.}

\item{x}{(character string) polynomial variable or process variable.}

\item{t}{(character string) time/index variable.}
}
\value{
character string.
}
\description{
TeX Matrix Polynomial Filters
}
\examples{
coefs = array(round(rnorm(2*3*1), 1), dim = c(2,3,2))

as_tex_matrixfilter(coefs, x = '\\\\epsilon', t = 's')
}
