% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_validation.R
\name{check objects}
\alias{check objects}
\alias{is.stable}
\alias{is.stable.ratm}
\alias{is.miniphase}
\alias{is.miniphase.ratm}
\title{Check Properties of Rational Matrices}
\usage{
is.stable(x, ...)

\method{is.stable}{ratm}(x, ...)

is.miniphase(x, ...)

\method{is.miniphase}{ratm}(x, ...)
}
\arguments{
\item{x}{a rational matrix object, i.e. a \code{\link{polm}}, \code{\link{lmfd}}, 
\code{\link{stsp}}, \code{\link{pseries}}, or \code{\link{zvalues}} object.}

\item{...}{not used.}
}
\value{
Boolean. Note \code{is.stable}, \code{is.miniphase} return \code{NA} 
        for objects of type \code{pseries} and \code{zvalues}.
}
\description{
Check Properties of Rational Matrices
}
\examples{
a = test_polm(dim = c(2,2), deg = 1, random = TRUE)
b = test_polm(dim = c(2,3), deg = 1, random = TRUE)
c = lmfd(a,b)

is.stable(a)
is.stable(c)
is.stable(as.stsp(c))
is.stable(pseries(c))

is.miniphase(b[,1:2])
is.miniphase(as.stsp(c)[,1:2])

is.miniphase(b)
is.miniphase(c)
}
