% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_polynomial_methods.R
\name{companion_matrix}
\alias{companion_matrix}
\title{Companion Matrix of a Polynomial Matrix}
\usage{
companion_matrix(a)
}
\arguments{
\item{a}{A square polynomial matrix, i.e. an object of class \code{\link{polm}}.}
}
\value{
A (companion) matrix of dimensions \eqn{(mp,mp)}.
}
\description{
Computes a companion matrix for a square (\eqn{m,m)}-dimensional), matrix polynomial 
\deqn{a(z) = a_0 + a_1 z + \cdots + a_p z^p}{a(z) = a[0] + a[1] z + \dots + a[p] z^p}
The companion matrix is e.g. used to determine the zeroes of a polynomial matrix,
see \code{\link{zeroes}}. 
\cr
Note that the function throws an error, if the constant term \eqn{a_0}{a[0]} is singular.
There is no check whether some of the leading coefficients are zero. 
So the results is an \eqn{(mp,mp)}-dimensional matrix, even if \eqn{a_p}{a[p]} is zero.
}
\examples{
companion_matrix(polm(c(1,0,0,0.5,0))) # scalar polynomial
companion_matrix(polm(diag(3)))        # zero degree polynomial 
companion_matrix(polm(dbind(d = 3, diag(2), -test_array(dim = c(2,2,1)))))
companion_matrix(polm(dbind(d = 3, diag(2), -test_array(dim = c(2,2,2)))))

# the following examples throw an error
companion_matrix(polm(c(0,0,0,0.5))) # constant term is zero
companion_matrix(polm(test_array(dim = c(2,1,3)))) # non-square polynomial
companion_matrix(polm(test_array(dim = c(2,2,0)))) # zero polynomial
}
