% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_statespace_methods.R
\name{ctr_matrix}
\alias{ctr_matrix}
\alias{obs_matrix}
\title{Controllability and Observability Matrix}
\usage{
ctr_matrix(A, B, o = NULL)

obs_matrix(A, C, o = NULL)
}
\arguments{
\item{A}{either a \code{\link{stsp}} object or a square \eqn{(s,s)} dimensional matrix.}

\item{B}{\eqn{(s,n)} dimensional matrix. This argument is ignored if \code{A} is a 
\code{stsp} object.}

\item{o}{(non negative) integer. The default value is \eqn{o=s}.}

\item{C}{\eqn{(m,s)} dimensional matrix. This argument is ignored if \code{A} is a 
\code{stsp} object.}
}
\value{
Controllability or observability matrix.
}
\description{
The controllability matrix of a statespace realisation \eqn{k(z)=C(I-Az^{-1})^{-1}B + D} is 
the matrix
\deqn{
[B,AB,\dots,A^{o-1}B]
}
and the observability matrix is 
\deqn{
[C',A'C',\dots,(A')^{o-1}C']'
}
}
