% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_dim.R
\name{dim methods}
\alias{dim methods}
\alias{dim.polm}
\alias{dim.lpolm}
\alias{dim.lmfd}
\alias{dim.rmfd}
\alias{dim.stsp}
\alias{dim.pseries}
\alias{dim.zvalues}
\title{Dimensions of Objects}
\usage{
\method{dim}{polm}(x)

\method{dim}{lpolm}(x)

\method{dim}{lmfd}(x)

\method{dim}{rmfd}(x)

\method{dim}{stsp}(x)

\method{dim}{pseries}(x)

\method{dim}{zvalues}(x)
}
\arguments{
\item{x}{Object of type \code{\link{polm}}, \code{\link{lmfd}}, 
\code{\link{rmfd}}, \code{\link{stsp}}, \code{\link{pseries}}, 
or \code{\link{zvalues}}.}
}
\value{
Returns a named vector of integers (\code{m,n} always refer to the number of 
        rows and columns of the rational matrix)
        
\item{\code{c(m,n,p,min_deg}}{for Laurent polynomial matrices 
  (i.e. for \code{\link{lpolm}} objects).}
\item{\code{c(m,n,p)}}{for polynomial matrices 
  (i.e. for \code{\link{polm}} objects).}
\item{\code{c(m,n,p,q)}}{for left matrix fraction descriptions 
  (i.e. for \code{\link{lmfd}} objects).}
\item{\code{c(m,n,p,q)}}{for right matrix fraction descriptions 
  (i.e. for \code{\link{rmfd}} objects).}
\item{\code{c(m,n,s)}}{for state space representations 
  (i.e. for \code{\link{stsp}} objects).}
\item{\code{c(m,n,lag.max)}}{for power series expansions  
  (i.e. for \code{\link{pseries}} objects).}
\item{\code{c(m,n,n.f)}}{for frequency response functions 
  (i.e. for \code{\link{zvalues}} objects).}
}
\description{
Retrieve the dimension and degrees of rational matrix objects. 
E.g. for a polynomial matrix \eqn{x(z)} (i.e. for a \code{\link{polm}} object),
\code{dim(x)} returns the vector  \code{c(m,n,p)}, 
where \code{m,n} are the respective number of rows and columns of the matrix and \code{p} is the (polynomial) degree of the matrix. 
For a Laurent polynomial, there is an additional element in the vector \code{c(m,n,p,min_deg)} returned by \code{dim(x)} pertaining to the (possibly negative) minimal degree. 
For a rational matrix in RMFD form \eqn{x(z)=d(z)c^{-1}(z)} (i.e. for an \code{\link{rmfd}} object),
\code{dim(x)} returns the vector \code{c(m,n,p,q)},
where \code{m,n} are the respective number of rows and columns of the matrix and 
\code{p,q} are the (polynomial) degrees of the polynomial matrices \eqn{c(z)} and \eqn{d(z)} respectively.
}
