% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_extract.R
\name{extract_matrix_}
\alias{extract_matrix_}
\title{Helper for Extracting indices}
\usage{
extract_matrix_(m, n, n_args, is_missing, i, j)
}
\arguments{
\item{m}{Integer. Output dimension.}

\item{n}{Integer. Input dimension.}

\item{n_args}{Integer. One, or two. Corresponds to the case where one set of indices or two sets are given.
Corresponds to x[i] and x[i,j]}

\item{is_missing}{Boolean vector of dimension \code{n_args}, i.e. one or two. 
TRUE when this function is used fo x[] and x[,]}

\item{i}{First set of indices. Can be integers or booleans.}

\item{j}{Second set of indices. Like above.}
}
\value{
Indices (integers) which will be further used to extract the relevant elements
}
\description{
Linear indices for extracting elements of a matrix are obtained, 
depending on the dimension of the matrix, number of arguments \code{n_args}, the missingness \code{is_missing} of these arguments, and their contents \code{i} and \code{j}.
}
\details{
No input checks are performed because this is an internal function. 
Consequently, some input combination are contradictory without alerting the user.
}
\examples{
# x[,]
extract_matrix_(6, 5, 
                n_args = 2,
                is_missing = c(TRUE, TRUE),
                i = 1:3,
                j = c(2,5))
                
# x[i,j]
extract_matrix_(6, 5, 
                n_args = 2,
                is_missing = c(FALSE, FALSE),
                i = 1:3,
                j = c(2,5))
                
# x[,j]
extract_matrix_(6, 5, 
                n_args = 2,
                is_missing = c(TRUE, FALSE),
                i = 1:3,
                j = c(2,5))
                
# x[i,]
extract_matrix_(6, 5, 
                n_args = 2,
                is_missing = c(FALSE, TRUE),
                i = 1:3,
                j = c(2,5))
                
# x[i], j is ignored if available (doesn't happen because it will not be called in this way)
extract_matrix_(6, 5, 
                n_args = 1,
                is_missing = c(FALSE, TRUE),
                i = 1:10,
                j = c(2,5))
                
# x[i], j is ignored if available (doesn't happen because it will not be called in this way)
extract_matrix_(6, 5, 
                n_args = 1,
                is_missing = c(FALSE, TRUE),
                i = c(3,6,7,1),
                j = c(2,5))
}
\keyword{internal}
