% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsp_methods.R
\name{grammians}
\alias{grammians}
\title{Grammians}
\usage{
grammians(
  obj,
  which = c("lyapunov", "minimum phase", "ctr", "obs", "obs_inv", "ctr_inv")
)
}
\arguments{
\item{obj}{(\code{\link{stsp}} object) rational matrix in statespace form.}

\item{which}{(character string) specifies the type of Grammian(s) to be computed.See 
below for more details.}
}
\value{
Either the selected Grammian (if \code{which} is one of \code{'ctr', 'obs', 'ctr_inv', 'obs_inv'}) 
        or a list with two components \code{P} and \code{Q} (for the case \code{which = 'lyapunov'} 
        or \code{which = 'miniumum phase'}).
}
\description{
The procedure computes "grammians" of a statespace realization, which may e.g. be used for 
balancing the statespace realization.
}
\details{
The \emph{controllability Grammian} \eqn{P} of a (stable) statespace realization 
\deqn{K(z) = C(Iz^{-1} - A)^{-1}B + D} 
is the solution of 
the Lyapunov equation \eqn{P = APA' + BB'}. The \emph{observability Grammian} is the solution 
of the Lyapunov equation \eqn{Q = A'QA + C'C}. If the statespace realization is \emph{stable} 
(the moduli of the eigenvalues of \code{A} are less than one) then \eqn{P,Q} are positive 
semidefinite 
and \eqn{P} is non singular if and only if the statespace realization is \emph{controllable} 
and \eqn{Q} is non singular if and only if the statespace realization is \emph{observable}. 
Hence the grammians may also be used to check whether the statespace realization is minimal 
(controllable \emph{and} observable). 

If the rational matrix is (strictly) minimum phase (i.e. \eqn{K(z)} is a square, invertible 
matrix and the eigenvalues of the matrix \eqn{(A - BD^{-1}C)} have moduli less than one) 
then we may also compute the controllability and the observability Grammian 
of the statespace realization 
\deqn{K^{-1}(z) = -D^{-1}C (Iz^{-1} - (A - BD^{-1}C))^{-1}BD^{-1} + D^{-1}.} 
of the inverse matrix \eqn{K^{-1}(z)}. These grammians have a similar interpretation. 

The above described grammians may be selected by setting the parameter \code{which} to 
\code{'ctr'}, \code{'obs'}, \code{'ctr_in'} or \code{'obs_inv'} respectively. 

For \emph{balancing} a statespace realization one needs a suitable pair of grammians. 
Two popular choices have been implemented: For \code{which = 'lyapunov'} the procedure 
returns the controllability and the observability Grammian and for \code{which = 'minimum phase'} 
the controllability matrix of the system and the observability Grammian of the inverse system 
are returned.

The procedure throws an error if the state space realization is not stable, respectively not 
minimum phase.
}
\examples{
# create a random, (3 by 2) rational matrix, 
# with a stable and minimum phase statespüace realization
obj = test_stsp(dim = c(3,2), s = 5, bpoles = 1, bzeroes = 1)
gr = grammians(obj, which = 'lyapunov')
gr

# we could also compute these grammians seperately 
all.equal(gr$P, grammians(obj,'ctr'))
all.equal(gr$Q, grammians(obj,'obs'))

# create a random (3 by 3) rational matrix, 
# with a stable and minimum phase statespüace realization
# Note: for the choice "minimum phase" the rational matrix 
# must be square and invertible.
obj = test_stsp(dim = c(3,3), s = 5, bpoles = 1, bzeroes = 1)
gr = grammians(obj, which = 'minimum phase')
gr

# we could also compute these grammians seperately 
all.equal(gr$P, grammians(obj,'ctr'))
all.equal(gr$Q, grammians(obj,'obs_inv'))
}
\seealso{
\code{\link{ctr_matrix}}, \code{\link{obs_matrix}}, \code{\link{lyapunov}} 
         and \code{\link{balance}}.
}
