% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_realization_tools.R
\name{hankel2nu}
\alias{hankel2nu}
\title{Compute left-Kronecker Indices for given Hankel Matrix}
\usage{
hankel2nu(H, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{H}{Block Hankel matrix, as computed e.g. by \code{\link{pseries2hankel}}.}

\item{tol}{tolerance parameter, used by \code{\link{qr}}.}
}
\value{
Integer vector with the Kronecker indices.
}
\description{
This (internal) helper functions determines the left-Kronecker indices 
given the Hankel matrix of the impulse response coefficients. 
There are no checks on parameters! In particular, note that 
the Hankel matrix must have an attribute \code{order=c(m,n,f,p)} 
which describes the block size \eqn{(m,n)} and the number 
of block rows (\eqn{f}) and block columns (\eqn{p}).
}
