% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_arithmetic_transpose.R
\name{Ht}
\alias{Ht}
\alias{Hermitean-transpose}
\alias{Ht.polm}
\alias{Ht.lpolm}
\alias{Ht.stsp}
\alias{Ht.zvalues}
\title{Hermitean Transpose}
\usage{
Ht(x)

\method{Ht}{polm}(x)

\method{Ht}{lpolm}(x)

\method{Ht}{stsp}(x)

\method{Ht}{zvalues}(x)
}
\arguments{
\item{x}{rational matrix object of class \code{\link{polm}}, \code{\link{lpolm}}, 
\code{\link{stsp}} or \code{\link{zvalues}}.}
}
\value{
A rational matrix object which represents the Hermitean transpose \eqn{x^*(z)}. 
        The Hermitean transpose of a polynomial matrix (in general) is a Laurent polynomial, 
        therefore the output is of class \code{lpolm} if \code{x} is a \code{polm} object. 
        In all other cases the output is of the same class as the input \code{x}.
}
\description{
The \emph{Hermitean transpose} of a rational matrix \eqn{x(z)} is defined as 
\deqn{x^{*}(z)=\overline{x(\bar{z}^{-1})}'.}{x^*(z)=Conj(x(Conj(z)^{-1}))'.} 
This means e.g. for a polynomial with real coefficients 
\eqn{x(z)=a_0 + a_1 z +\cdots + a_p z^p}{x(z)=a[0] + a[1] z +\dots + a[p] z^p} 
that the coefficient matrices are transposed and that \eqn{z} is replaced 
by \eqn{z^{-1}}: 
\eqn{x^*(z)=a_0' + a_1' z^{-1} +\cdots + a_p' z^{-p}}{
     x^*(z)=a[0]' + a[1]' z^{-1} +\dots + a[p]' z^{-p}}, i.e. 
the result is (in general) a Laurent polynomial.
}
\details{
The Hermitean transpose is only implemented for 
polynomial matrices (\code{\link{polm}} objects), 
Laurent polynomials \code{\link{lpolm}} objects), 
rational matrices in statespace form (\code{\link{stsp}} objects) and for 
the frequency response of rational matrices (\code{\link{zvalues}} objects).  

The case of rational matrices in LMFD or RMFD form has not (yet) been implemented. 
It is not possible to directly construct the Hermitean transpose from 
given power series coefficients. Therefore \code{Ht()} does 
not support \code{\link{pseries}} objects.

Finally note that the Hermitean transpose of a rational matrix 
\eqn{x(z)=D+zB(I-Az)^{-1}B} in statespace form 
(in general) has a pole at zero, if the state transition matrix \eqn{A} 
is singular. Since rational matrices with a pole at \eqn{z=0} have \bold{no} 
statespace realization, the procedure \code{Ht()} throws an error in this case.
}
\examples{
# rational matrix in statespace form 
x = test_stsp(dim = c(4,4), s = 2)
Ht(x)
all.equal(zvalues(Ht(x)), Ht(zvalues(x)))

# Note (x(z)  x^*(z)) is Hermitean for |z| = 1
xx = zvalues(x) \%r\% Ht(zvalues(x))
# print(xx, digits = 2)
apply(unclass(xx), MARGIN = 3, FUN = isSymmetric)

# polynomial matrix 
x = test_polm(dim = c(2,3), degree = 1)
Ht(x)

\dontrun{
# Ht is not implemented for lmfd and rmfd objects:
Ht(test_lmfd(dim = c(2,2), degrees = c(3,3)))
Ht(test_rmfd(dim = c(2,2), degrees = c(3,3)))
}
# Workaround: convert to pseries first
Ht(pseries(test_lmfd(dim = c(2,2), degrees = c(3,3))))

}
