% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_arithmetic_operations.R
\name{inflate_object}
\alias{inflate_object}
\title{Inflate a scalar to an (m x n) matrix with identical entries}
\usage{
inflate_object(e, m, n)
}
\arguments{
\item{e}{Object of type \code{\link{polm}}, \code{\link{lpolm}},  \code{\link{stsp}}, \code{\link{pseries}}, or \code{\link{zvalues}}}

\item{m}{Integer. Output dimension to which the scalar should be inflated.}

\item{n}{Integer. Output dimension to which the scalar should be inflated.}
}
\value{
Object of same type but with *inflated* dimension
}
\description{
Used in elementwise group operations \code{\link{Ops.ratm}} to *inflate* a scalar \code{ratm} object such that it can be, e.g., elementwise added.
It is used after \code{\link{upgrade_objects}} is called and therefore only relevant for 
\itemize{
\item \code{\link{polm}}, 
\item \code{\link{lpolm}}, 
\item \code{\link{stsp}}, 
\item \code{\link{pseries}},
\item \code{\link{zvalues}}
}
In other words, one cannot inflate \code{\link{lmfd}} or \code{\link{rmfd}} objects.
}
\examples{
(lp = test_lpolm(degree_max = 1, degree_min = -1))
inflate_object(lp, 2, 2)

(ss = test_stsp(s = 1))
inflate_object(ss, 2, 2)
}
