% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsp_methods.R
\name{is.minimal}
\alias{is.minimal}
\alias{is.minimal.stsp}
\title{Check the Minimality of a Statespace Realization}
\usage{
is.minimal(x, ...)

\method{is.minimal}{stsp}(x, tol = sqrt(.Machine$double.eps), only.answer = TRUE, ...)
}
\arguments{
\item{x}{(\code{\link{stsp}} object) rational matrix in statespace form.}

\item{...}{not used.}

\item{tol}{a tolerance parameter, which is used to decide the rank of 
the Hankel matrix of the impulse response coefficients.}

\item{only.answer}{if TRUE, just return a logical (TRUE or FALSE). 
Otherwise a list with additional information is returned.}
}
\value{
If \code{only.answer} is true then a logical (\code{TRUE} or \code{FALSE}) is returned.
  Otherwise, a list with the following slots is returned. 
  \item{answer}{A boolean as above.}
  \item{H}{Hankel matrix of impulse response coefficients (with s block rows and s block columns).}
  \item{sv}{The singular values of H.}
  \item{s0}{(integer) (estimate of the) rank of H, i.e. the minimal statespace dimension.}
}
\description{
Check whether a statespace realization of a rational matrix is minimal. 
The procedure constructs the Hankel matrix of the impulse response coefficients with \eqn{s} block rows and \eqn{s} block columns, 
where \eqn{s} is the statespace dimension of the given statespace realization. 
If this statespace realization is minimal then the Hankel matrix has rank \eqn{s}. 
Therefore the procedure returns \code{TRUE} if the \eqn{s}-th singular values of the Hankel matrix is larger than \code{tol}.
}
\details{
The procedure does not check whether the statespace realization is \emph{observable} and/or 
\emph{controllable}. To this end one may compute the observability/controllability matrices (\code{\link{obs_matrix}}, \code{\link{ctr_matrix}}) or the corresponding Grammians (\code{\link{grammians}}).
}
\note{
This procedure returns different objects, depending on the parameter \code{only.answer}.
}
\examples{
x = test_stsp(dim = c(2,2), s = 2)
is.minimal(x)
# note that operations on "stsp" objects may return non minimal realizations
# 
is.minimal(rbind(x, x), only.answer = FALSE)[c('answer','sv','s0')]
is.minimal(x \%r\% (x^(-1)), only.answer = FALSE)[c('answer','sv','s0')]

is.minimal(test_stsp(dim = c(2,0), s = 2))
is.minimal(test_stsp(dim = c(0,2), s = 0))
}
\seealso{
\code{\link{pseries}} and \code{\link{bhankel}}.
}
