% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_validation.R
\name{is.class methods}
\alias{is.class methods}
\alias{is.polm}
\alias{is.lpolm}
\alias{is.lmfd}
\alias{is.rmfd}
\alias{is.stsp}
\alias{is.pseries}
\alias{is.zvalues}
\title{Check Objects}
\usage{
is.polm(x)

is.lpolm(x)

is.lmfd(x)

is.rmfd(x)

is.stsp(x)

is.pseries(x)

is.zvalues(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{TRUE} if the argument is a valid object.
}
\description{
Check if the argument is a valid object.
}
