% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representation_classes.R
\name{lmfd}
\alias{lmfd}
\title{Constructor for Left Matrix Fraction Descriptions (LMFDs)}
\usage{
lmfd(a, b)
}
\arguments{
\item{a, b}{\code{\link{polm}} objects, or objects which may be coerced to a \code{\link{polm}} object,
via \code{x = polm(x)}. Either of the two arguments may be omitted.}
}
\value{
An object of class \code{lmfd}.
}
\description{
A Left Matrix Fraction Description (LMFD) of a rational matrix, \eqn{x(z)} say, is a
pair \eqn{(a(z),b(z))} of polynomial matrices, such
that \eqn{x(z) = a^{-1}(z) b(z)}. 
The polynomial matrix \eqn{a(z)} must be square and invertible.
}
\details{
Suppose that \eqn{x(z)=a^{-1}(z) b(z)} is an \eqn{(m,n)}-dimensional matrix and that 
\eqn{a(z)} and \eqn{b(z)} have degrees \eqn{p} and \eqn{q} respectively. 
The corresponding \code{lmfd} object stores the coefficients of the polynomials \eqn{a(z), b(z)} in 
an \eqn{(m,m(p+1)+n(q+1))} dimensional (real or complex valued) matrix together with an 
attribute \code{order = c(m,n,p,q)} and a class attribute \code{c("lmfd", "ratm")}. 

For a valid LMFD we require \eqn{m>0} and \eqn{p\geq 0}{p \ge 0}.
}
\examples{
### (1 x 1) rational matrix x(z) = (1+z+z^2)^(-1) (3+2z+z^2)
lmfd(c(1,1,1), c(3,2,1)) \%>\% print(format = 'c')

### (1 x 1) rational matrix x(z) = (3+2z+z^2)
lmfd(b = c(3,2,1)) \%>\% print(format = 'c')

### (1 x 1) rational matrix x(z) = (1+z+z^2)^(-1)
lmfd(c(1,1,1)) \%>\% print(format = 'c')

### (2 x 3) rational matrix with degrees p=1, q=1
x = lmfd(array(rnorm(2*2*2), dim = c(2,2,2)), 
         array(rnorm(2*3*2), dim = c(2,3,2)))
is.lmfd(x)
dim(x)
str(x)
print(x, digits = 2)

### the following calls to lmfd() throw an error
lmfd() # no arguments!
lmfd(a = test_polm(dim = c(2,3), degree = 1))  # a(z) must be square
lmfd(a = test_polm(dim = c(2,2), degree = -1)) # a(z) must have degree >= 0
}
\seealso{
Useful methods and functions for the \code{lmfd} class are: 
\itemize{
   \item \code{\link{test_lmfd}} generates random rational matrices in LMFD form.
   \item checks: \code{\link{is.lmfd}}, \code{\link{is.miniphase}}, \code{\link{is.stable}} 
         and \code{\link{is.coprime}}.
   \item generic S3 methods: \code{\link[rationalmatrices]{dim}}, 
         \code{\link[rationalmatrices]{str}} and \code{\link[rationalmatrices]{print}}.
   \item arithmetics: \code{\link[rationalmatrices]{Ops.ratm}}.
   \item matrix operations: \code{\link[rationalmatrices]{bind}}.
   \item extract the factors \eqn{a(z)} and \eqn{b(z)} with \code{\link{$.lmfd}}.      
   \item \code{\link{zeroes}}, \code{\link{pseries}}, \code{\link{zvalues}}, ...  
}
}
