% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_reflection_poles_zeroes.R
\name{make_allpass}
\alias{make_allpass}
\title{Create an All-Pass Rational Matrix}
\usage{
make_allpass(A, B)
}
\arguments{
\item{A}{square \eqn{(s,s)} matrix}

\item{B}{\eqn{(s,m)} matrix}
}
\value{
A \code{\link{stsp}} object which represents the all-pass 
        rational matrix.
}
\description{
Create a square, all-pass rational matrix in statespace form with given \eqn{A,B}.
}
\section{Notes}{
 
The function is intended as an internal helper function and thus does not check the inputs. See also \code{\link{reflect_poles}} and \code{\link{reflect_zeroes}}. 

If \eqn{A,C} is given we can use \code{t(make_allpass(t(A), t(C)))}.
}

\examples{
m = 2
s = 6
A = matrix(rnorm(s*s), nrow = s, ncol = s)
B = matrix(rnorm(s*m), nrow = s, ncol = m)

K = make_allpass(A, B)
all.equal(cbind(A,B), cbind(K$A, K$B))
# check that K(z) is all-pass
Kf = zvalues(K)
all.equal(zvalues(polm(diag(m))) + complex(imaginary = 0), 
          Kf \%r\% Ht(Kf) + complex(imaginary = 0))

# if (A,C) is given, we proceed as follows:
C = matrix(stats::rnorm(m*s), nrow = m, ncol = s)
K = t(make_allpass(t(A), t(C)))
all.equal(rbind(A,C), rbind(K$A, K$C))
# check that K(z) is all-pass
Kf = zvalues(K)
all.equal(zvalues(polm(diag(m))) + complex(imaginary = 0), 
          Kf \%r\% Ht(Kf) + complex(imaginary = 0))
}
