% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_visualization_tools.R
\name{plot_axes}
\alias{plot_axes}
\title{Plot Grid Lines, Axes and Axes Labels}
\usage{
plot_axes(
  axes = c(TRUE, TRUE, FALSE, FALSE),
  tick_labels = axes,
  x_date = 0,
  titles = rep(NA_character_, 4),
  style = style_parameters("bw"),
  parse_titles = FALSE
)
}
\arguments{
\item{axes}{4-dimensional boolean vector, indicating whether an axis should be put
at the respective side of the plot.}

\item{tick_labels}{4-dimensional boolean vector, indicating whether the respective
axis has tick-labels.}

\item{x_date}{scalar number, used to convert the (numeric) x-axis limits \code{xlim} to
a date/time object. If \code{x_date} is of class \code{POSIXct}
then the limits are converted as \code{as.POSIXct(xlim, origin = x_date)}.
If \code{x_date} is of class \code{Date}
then the limits are converted as \code{as.Date(xlim, origin = x_date)}.
Otherwise no coercion is performed. Finally
\code{graphics::Axis(x = xlim, ...)} is called to do the actual plotting
of the x-axis.}

\item{titles}{a 4-dimensional character vector. If the respective entry is \code{NA} then
no label/title is put at this side of the plot.}

\item{style}{list of "style" parameters as returned by \code{\link{style_parameters}}.}

\item{parse_titles}{if TRUE the titles are coerced to \code{\link{expression}} before
plotting.}
}
\value{
(invisible) \code{NULL}.
}
\description{
This tools plots the grid lines,axes and axes titles/labels. The x-axis may be
a date/time axis.
}
\examples{
date = seq(as.Date("2019/1/1"), by = "month", length.out = 24)
titles = c(NA, 'y','main', NA)
plot(date, 1:24, axes = FALSE, xlab = NA, ylab = NA)
plot_axes(axes = c(TRUE, FALSE, FALSE, TRUE), titles = titles,
          style = style_parameters('gray'), x_date = date[1] - as.numeric(date[1]))
points(date, 1:24)
}
\keyword{internal}
