% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polm_methods.R
\name{polm2fwd}
\alias{polm2fwd}
\title{Transforms to Polynomial in Forward Shift}
\usage{
polm2fwd(polm_obj)
}
\arguments{
\item{polm_obj}{\code{\link{polm}} object}
}
\value{
\code{\link{lpolm}} object
}
\description{
Transform \code{\link{polm}} object to polynomial in forward shift (represented as \code{\link{lpolm}} object), i.e. transform
\deqn{a(z) = a_0 + a_1 z^1 + \cdots + a_p z^p}{
      a(z) = a[0] + a[1] z^1 + \dots + a[p] z^p}
to       
\deqn{a(z) = a_0 + a_1 z^{-1} + \cdots + a_p z^{-p}}{
      a(z) = a[0] + a[1] z^(-1) + \dots + a[p] z^(-p)}
}
\examples{
(p = test_polm(degree = 3))
polm2fwd(p)
}
