% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect_poles_zeroes.R
\name{polm_div}
\alias{polm_div}
\title{Division Algorithm for Polynomial Matrices}
\usage{
polm_div(a, b)
}
\arguments{
\item{a, b}{Two compatible polynomial matrices.}
}
\value{
List with two slots 
\item{qu}{contains the polynomial \eqn{c(z)}}
\item{rem}{contains the polynomial \eqn{d(z)}.}
}
\description{
For given polynomial matrices \eqn{a(z), b(z)} compute two matrices 
\eqn{c(z), d(z)} such that 
\deqn{a(z) = c(z) b(z) + d(z)}
where the degree of \eqn{d(z)} is smaller than the degree of \eqn{b(z)}. 
The matrix \eqn{b(z)} must be square with a non singular leading coefficient 
matrix! The matrices must be compatible, i.e. the number of columns of \eqn{a(z)} 
must equal the number of rows (and columns) of \eqn{b(z)}.
}
\examples{
a = test_polm(dim = c(3,2), degree = 4, random = TRUE)
b = test_polm(dim = c(2,2), degree = 2, random = TRUE)
(out = polm_div(a, b))
all.equal(a, out$qu \%r\% b + out$rem)
}
