% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_3D}
\alias{print_3D}
\title{Helper for Printing Matrix (Laurent) Polynomials}
\usage{
print_3D(
  a,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  laurent = FALSE
)
}
\arguments{
\item{a}{Array with 3 dimensions, representing a polynomial or Laurent polynomial}

\item{digits}{Default NULL. Otherwise an integer specifying the number of digits of the coefficients that should be printed, 
see \code{\link{round}}.}

\item{format}{Default "i|jz", i.e. the matrix polynomial is printed that the "z part is the slowest moving index.
The vertical bar designates separation across array dimensions (one bar = two array dimensions, two bars = array of dimension 3).
In addition to multiple ordering options regarding row-index \code{i}, column-index \code{j}, and polynomial power \code{z}, 
there is also the option \code{character} (partial matching is enabled such that \code{format = "c"} gives the desired result)
which prints each univariate (Laurent) polynomial into a matrix with appropriate numbers of rows and columns.
Note that \code{\link{stsp}} objects have no format option. 
The option \code{'character'} is only implemented for polynomials, Laurent polynomials, LMFDs and RMFDs with real coefficients.}

\item{laurent}{Boolean or integer. Default set to FALSE.
If one deals with Laurent polynomials, an integer corresponding to the minimal degree of the Laurent polynomial should be supplied.}
}
\value{
Printed (Laurent) polynomial matrix
}
\description{
Called by the \code{rationalmatrices} methods for the \link{print} S3 generic.
}
\keyword{internal}
