% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representation_conversions.R
\name{pseries2lmfd}
\alias{pseries2lmfd}
\title{Construct a LMFD Representation from Impulse Response}
\usage{
pseries2lmfd(obj, Hsize = NULL, nu = NULL, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{obj}{\code{\link{pseries}} object or 3-D array with dimension \eqn{(m,n,l+1)}.}

\item{Hsize}{integer vector \code{c(f,p)}, number of block rows and block columns
of the Hankel matrix which is used to construct the LMFD.
If NULL a default choice is made.}

\item{nu}{integer vector with the Kronecker indices. If \code{NULL} then the
Kronecker indices are computed via a QR decomposition of the
transpose of the Hankel matrix, see \code{\link{qr}}.}

\item{tol}{tolerance parameter, used by \code{\link{qr}}.}
}
\value{
List with components
\item{Xl}{\code{\link{lmfd}} object which contains the LMFD representation
         of the rational matrix (in echelon form).}
\item{nu}{integer vector with the Kronecker indices.}
}
\description{
This (helper) function constructs a left matrix fraction description
of a rational matrix from its impulse response. 
Of course the impulse response must contain sufficiently many lags. 
The constructed LMFD is in \emph{echelon canonical} form.
}
\details{
There are a number of restrictions on the dimension \eqn{(m,n)}
and number of lags \eqn{l} of the impulse response,
the number of block rows (\eqn{f}), block columns (\eqn{p})
of the Hankel matrix and the Kronecker indices \eqn{\nu_i}{\nu[i]}:

We require that:
\eqn{m>0}, \eqn{p>0}, \eqn{f>1},
\eqn{l \geq f+p-1}{l\ge f+p-1} and \eqn{\nu_i <f}{\nu[i] < f}.
If these restrictions are not satisfied an error is thrown.
}
\examples{
# generate a random LMFD object
m = 3
n = 2
p = 1
q = 1
a = test_polm(dim = c(m,m), degree = p, random = TRUE, bzeroes = 1)
b = test_polm(dim = c(m,n), degree = q, random = TRUE, bzeroes = 1)
Xl = lmfd(a,b)

# compute impulse response of this matrix
Xi = pseries(Xl, lag.max = 2*(max(p,q)+1))

# reconstruct a matrix from this impulse response
out = pseries2lmfd(Xi)
out

# check that the lmfd object is a realization of the given impulse response
Xi1 = pseries(out$Xl, lag.max = 2*(max(p,q)+1))
all.equal(Xi, Xi1)
}
