% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_polynomial_methods.R
\name{snf}
\alias{snf}
\title{Smith Normal Form}
\usage{
snf(a, tol = sqrt(.Machine$double.eps), debug = FALSE)
}
\arguments{
\item{a}{Polynomial matrix, i.e. an object  of class \code{\link{polm}}.}

\item{tol}{Tolerance parameter, used for "pruning" the polynomial matrix (after each step). 
See \code{\link{prune}}.}

\item{debug}{Logical, default to FALSE. If TRUE, then some diagnostic messages are printed.}
}
\value{
A list with five elements
\item{\code{s}}{Matrix polynomial of class \code{\link{polm}} and dimensions \eqn{(m,n)},
    representing \eqn{s(z)} in the Smith-form (whose only non-zero elements are on the diagonal).}
\item{\code{u}}{Unimodular matrix polynomial of class \code{\link{polm}} and dimensions \eqn{(m,m)},
    Represents \eqn{u(z)} in the Smith form \eqn{a(z) = u(z) s(z) v(z)}.}
\item{\code{u_inv}}{Unimodular matrix polynomial of class \code{\link{polm}} and dimensions \eqn{(m,m)},
    Represents the inverse of \eqn{u(z)}.}
\item{\code{v}}{Unimodular matrix polynomial of class \code{\link{polm}} and dimensions \eqn{(n,n)},
    Represents \eqn{v(z)} in the Smith form \eqn{a(z) = u(z) s(z) v(z)}.}
\item{\code{v_inv}}{Unimodular matrix polynomial of class \code{\link{polm}} and dimensions \eqn{(n,n)},
    Represents the inverse of \eqn{v(z)}.}
}
\description{
Calculates the \emph{Smith normal form} of an \eqn{(m,n)}-dimensional polynomial matrix \eqn{a(z)},
i.e. a factorization of \eqn{a(z)} of the form 
\deqn{a(z) = u(z) s(z) v(z),}
where \eqn{u(z)} and \eqn{v(z)} are unimodular polynomial matrices of dimensions \eqn{(m,m)} and 
\eqn{(n,n)} respectively, and \eqn{s(z)} is a quasi-diagonal polynomial matrix of dimension \eqn{(m,n)}
with diagonal elements \eqn{d_i(z)}{d[i,i]} which satisfy
\itemize{
\item \eqn{d_{ii}}{d[i,i]} is monic for \eqn{i\leq r}{i \le r} and 
      \eqn{d_{ii}}{d[i,i]} is zero for \eqn{i>r}, and 
\item \eqn{d_{ii}}{d[i,i]} divides \eqn{d_{i+1,i+1}}{d[i+1,i+1]} for \eqn{i < r}.
}
Here \eqn{0\leq r \leq \min(m,n)}{0\le r \le min(m,n)} is the rank of \eqn{a(z)} 
when considered as a rational matrix.   
See *Gohberg, Lancaster, Rodman 09 - Matrix Polynomials* page 318 or \insertCite{Hannan.Deistler12}{rationalmatrices} page 42, Lemma 2.2.3.
}
\examples{
##############
# Quadratic case

a = test_polm(dim = c(2,2), degree = 1)
out = snf(a)

print(out$s, digits = 2, format = 'c')

all.equal(a, prune(out$u \%r\% out$s \%r\% out$v))

##############
# Tall case

a = test_polm(dim = c(3,2), degree = 1)
out = snf(a)

print(out$s, digits = 2, format = 'c')

all.equal(a, prune(out$u \%r\% out$s \%r\% out$v))

##############
# Wide case

a = test_polm(dim = c(2,3), degree = 1)
out = snf(a)

print(out$s, digits = 2, format = 'c')

all.equal(a, prune(out$u \%r\% out$s \%r\% out$v))

##############
# Diagonal case 
z = polm(c(0,1))
a = polm(diag(3))
a[3,3] = 1+2*z 
a[2,2] = a[3,3] * (1-z)
a[1,1] = a[2,2] * (1+z)
print(a, format = 'c')

out = snf(a)

print(out$s, digits = 2, format = 'c')

all.equal(a, prune(out$u \%r\% out$s \%r\% out$v))

##############
# Common factor(s) 
a = test_polm(dim = c(3,3), degree = 1, random = TRUE, digits = 1)
a = a * z
a[,2] = a[,2] * (1+z)
a[,1] = a[,2] * (1-z)

print(a, format = 'c')

out = snf(a)

print(out$s, digits = 2, format = 'c')

all.equal(a, prune(out$u \%r\% out$s \%r\% out$v))
}
\references{
\insertRef{Hannan.Deistler12}{rationalmatrices}
}
