% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representation_classes.R
\name{stsp}
\alias{stsp}
\title{Constructor for Statespace Realizations}
\usage{
stsp(A, B, C, D)
}
\arguments{
\item{A}{\eqn{(s,s)} matrix (or a vector of length \eqn{(s^2)})}

\item{B}{\eqn{(s,n)} matrix (or a vector of length \eqn{(sn)})}

\item{C}{\eqn{(m,s)} matrix (or a vector of length \eqn{(ms)})}

\item{D}{\eqn{(m,n)} matrix (or a vector of length \eqn{(mn)})}
}
\value{
An object of class \code{stsp}.
}
\description{
Any rational (\eqn{(m,n)}-dimensional) matrix \eqn{x(z)} which has no pole at 
\eqn{z=0} may be represented as 
\deqn{x(z) = C(z^{-1} I_s - A)^{-1} B + D}{x(z) = C(z^{-1} I - A)^{-1} B + D}
Here \eqn{I_s}{I} denotes the \eqn{(s,s)}-dimensional identity matrix and \eqn{A,B,C,D} 
are (real or complex valued) matrices of size \eqn{(s,s)}, \eqn{(s,n)}, \eqn{(m,s)} and \eqn{(m,n)}
respectively. The integer \eqn{s \geq 0}{s \ge 0} is called the \emph{state dimension} 
of the above \emph{statespace realization} of \eqn{x(z)}.
}
\details{
Internally statespace realizations are stored as an \eqn{(s+m, s+n)}-dimensional matrix 
with an attribute \code{order = c(m,n,s)} and a class attribute 
\code{c("stsp","ratm")}.

Any of the integers \eqn{m,n,s} may be zero. 

If the arguments \code{A,B,C} are missing, then a statespace realization with 
statespace dimension \eqn{s=0} is constructed. In this case \code{D} must a be matrix or 
a scalar. 
If \code{A,B,C} are given (and compatible) and \code{D} is missing, then 
\code{D = diag(x = 1, nrow = m, ncol = n)} is used.
}
\examples{
### x(z) =  I
stsp(D = diag(3))

### random (2 x 3) rational matrix in statespace form with state dimension s = 4
x = stsp(A = stats::rnorm(4*4), B = stats::rnorm(4*3), C = stats::rnorm(2*4))

is.stsp(x)
dim(x)
str(x)
print(x, digits = 3)
}
\seealso{
\itemize{
   \item \code{\link{test_stsp}} generates random polynomials. 
   \item checks: \code{\link{is.stsp}}, \code{\link{is.miniphase}}, 
         \code{\link{is.stable}} and \code{\link{is.minimal}}.
   \item generic S3 methods: \code{\link[rationalmatrices]{dim}}, 
         \code{\link[rationalmatrices]{str}} and \code{\link[rationalmatrices]{print}}.
   \item arithmetics: \code{\link[rationalmatrices]{Ops.ratm}}, matrix multiplication \code{\link{\%r\%}}, ... 
   \item matrix operations: transpose \code{\link[rationalmatrices]{t.stsp}}, Hermitean transpose 
         \code{\link[rationalmatrices]{Ht}}, \code{\link[rationalmatrices]{bind}}, 
         \code{\link{[.stsp}}, ...
   \item extract the system matrices \eqn{A,B,C,D} with \code{\link{$.stsp}}.      
   \item statespace realizations related tools: state transfromation \code{\link{state_trafo}}, 
         observability and controllability matrices \code{\link{obs_matrix}} and 
         \code{\link{ctr_matrix}},  Grammian matrices \code{\link{grammians}} and 
         balanced realizations \code{\link{balance}}.. 
   \item \code{\link{reflect_poles}} and \code{\link{reflect_zeroes}} may be used 
         to reflect poles and zeroes of a rational matrix in statespace form by multiplication
         with allpass rational matrices.
   \item \code{\link{zeroes}}, \code{\link{pseries}}, \code{\link{zvalues}}, ...  
}
}
