% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_tools.R
\name{test_array}
\alias{test_array}
\title{Create Test Array}
\usage{
test_array(dim, random = FALSE, dimnames = FALSE)
}
\arguments{
\item{dim}{Integer vector.}

\item{random}{Boolean. Default set to FALSE.
If TRUE, the elements are randomly generated (standard normal).}

\item{dimnames}{(logical) decides whether \code{dimnames} should be created.}
}
\value{
Real valued array of dimension \code{dim}.
}
\description{
The helper function \code{test_array} creates arrays of given dimension.
}
\details{
The function \code{test_array} creates an array with either 
\itemize{
\item entries of the form \code{x[i,j,...] = i*10^(n-1) + j*10^(n-2) + ...},
      where \code{n} is the number of dimensions, 
\item or randomly generated (standard normal) entries.
}

This function is mainly used to test operations on arrays
(like \code{\link{btoeplitz}}, \code{\link{bhankel}}, \code{\link{bmatrix}} and 
\code{\link{dbind}}).
}
\examples{
test_array(dim = 5)
test_array(dim = c(2,4), dimnames = TRUE)
test_array(dim = c(2,4,3), dimnames = FALSE)

# the examples below throw an error
test_array(dim = c())
test_array(dim = c(2,-1,2))
}
