% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_tools.R
\name{test_stsp}
\alias{test_stsp}
\title{Create Test Rational Matrix in Statespace Form}
\usage{
test_stsp(
  dim = c(1, 1),
  s = NULL,
  nu = NULL,
  D = NULL,
  digits = NULL,
  bpoles = NULL,
  bzeroes = NULL,
  n.trials = 100
)
}
\arguments{
\item{dim}{integer vector \code{c(m,n)}.}

\item{s}{state dimension (or NULL).}

\item{nu}{vector with the Kronecker indices (or \code{NULL}). Either the statespace dimension \code{s} 
or the Kronecker indices \code{nu} must be non \code{NULL}. If both parameters are given, 
then the parameter \code{s} is ignored.}

\item{D}{\eqn{(m,n)} dimensional matrix (or \code{NULL}). See the details below.}

\item{digits}{integer, if non NULL then the randomly generated numbers are rounded to 
"digits" number of decimal places.}

\item{bpoles}{lower bound for the moduli of the poles of the rational matrix (or NULL).}

\item{bzeroes}{lower bound for the moduli of the zeroes of the rational matrix (or NULL). 
This parameter is ignored for non-square matrices (m != n).}

\item{n.trials}{maximum number of trials.}
}
\value{
\code{\link{stsp}} object, which represents the generated rational matrix 
        \eqn{x(z)} in statespace form.
}
\description{
This simple tool may be used to create a random, \eqn{(m,n)}-dimensional, rational matrix
in statespace form with statespace dimension \eqn{s}.
}
\details{
If the Kronecker indices (parameter \code{nu}) 
are given, then a statespace model in echelon canonical form is generated. This means that 
some of the entries of the \eqn{A,B,C} matrices are fixed to be one or zero and the 
others are considerd as "free". See also \code{\link{Kronecker-Indices}}. 
The entries of the \eqn{A, B, C} matrices, which are not a priori fixed 
are randomly generated. 

If only the state dimension \eqn{s} (parameter \code{s}) 
is given, then all entries of the \eqn{A, B, C} matrices 
are considered as "free". 

The \eqn{D} matrix defaults to a \eqn{(m,n)}-dimensional diagonal matrix 
with ones on the diagonal (\code{diag(x=1, nrow = m, ncol = n)}). However, one may 
also pass an arbitray (compatible) \eqn{D} matrix to the procedure. 
This matrix may contain \code{NA}'s, which then are replaced by random numbers. 

The user may prescribe lower bounds for the moduli of the poles and the zeroes 
of the rational matrix. In this case the procedure simply generates (up to \code{n.trials}) 
random matrices until a matrix is found which satisfies the constraints. The standard deviation 
of the normal distribution, which is used to generate the random entries, is decreased in each step. 
Of course this is a very crude method and it may fail or need a very large number of randomly 
generated matrices.

Note also, that the generated model may be non-minimal.
}
\examples{
## random (2 x 2) statespace realization with state dimension s = 3
## no poles and zeroes within the unit circle
x = test_stsp(dim = c(2,2), s = 3, digits = 2, bpoles = 1, bzeroes = 1)
x
min(abs(poles(x))) > 1
min(abs(zeroes(x))) > 1
pseries2nu(pseries(x, lag.max = 5)) # Kronecker indices

## random (3 x 2) statespace realization in echelon canonical form
## D is lower triangular (with ones on the diagonal)
## no poles within the unit circle
x = test_stsp(dim = c(3, 2), nu = c(2,3,0), D = matrix(c(1,NA,NA,0,1,NA), nrow = 3, ncol = 2), 
              digits = 2, bpoles = 1)

x
min(abs(poles(x))) > 1
pseries2nu(pseries(x, lag.max = 10)) # check Kronecker indices
}
