% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_arithmetic_transpose.R
\name{transpose}
\alias{transpose}
\alias{t.polm}
\alias{t.lpolm}
\alias{t.lmfd}
\alias{t.rmfd}
\alias{t.stsp}
\alias{t.pseries}
\alias{t.zvalues}
\title{Rational Matrix Transpose}
\usage{
\method{t}{polm}(x)

\method{t}{lpolm}(x)

\method{t}{lmfd}(x)

\method{t}{rmfd}(x)

\method{t}{stsp}(x)

\method{t}{pseries}(x)

\method{t}{zvalues}(x)
}
\arguments{
\item{x}{rational matrix object, i.e. a \code{\link{polm}}, \code{\link{lpolm}}, 
\code{\link{lmfd}}, \code{\link{rmfd}}, \code{\link{stsp}}, 
\code{\link{pseries}} or \code{\link{zvalues}} object.}
}
\value{
A rational matrix object,
   which represents the transposed rational matrix \eqn{x'(z)}. 
   The output is of the same class as the input \code{x} unless 
   \code{x} is an \code{\link{lmfd}} or an \code{\link{rmfd}} object: 
   The transposition of 
   an \code{lmfd} object is an \code{rmfd} object, and vice versa.
}
\description{
Compute the transpose of a rational matrix \eqn{x(z)}.
}
\examples{
x = test_polm(dim = c(2,3), degree = 3, random = TRUE)
all.equal(pseries(t(x)), t(pseries(x)))

x = test_stsp(dim = c(3,2), s = 1)
all.equal(zvalues(t(x)), t(zvalues(x)))

# the transpose of an LMFD object is RMFD
(x = test_lmfd(dim = c(3,2), degrees = c(1,1)))
t(x)
all.equal(x, t(t(x)))

}
