% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_methods.R
\name{upgrade_objects}
\alias{upgrade_objects}
\title{Upgrade Objects to Common Class}
\usage{
upgrade_objects(force = TRUE, ...)
}
\arguments{
\item{force}{Default set to TRUE. 
If FALSE, then objects are also coerced when only one object is supplied.
Option FALSE is used in bind methods}

\item{...}{Arbitrary number of objects with superclass \code{ratm}}
}
\value{
Upgraded inputs
}
\description{
Used in \code{\link{cbind}}, 
\code{\link{rbind}}, 
\code{\link[rationalmatrices]{ratm_mult}}, 
and \code{\link{Ops.ratm}} 
to make operations on different subclasses of the (super-) class \code{ratm} possible.
}
\details{
Constant matrices are always upgraded to \code{\link{polm}},
and \code{\link{lmfd}} and \code{\link{rmfd}} are always upgraded to \code{\link{stsp}} if two \code{ratm} objects are involved.
Therefore, most operations are performed in the state space setting.

Depending on the highest "grade" of any object involved in the operation, the lower ranked object is upgraded to the highest one. 
The following ranks are assigned to the \code{ratm} objects:

\itemize{
  \item 1: \code{\link{polm}}
  \item 2: \code{\link{lpolm}}
  \item 3: \code{\link{lmfd}}
  \item 4: \code{\link{rmfd}}
  \item 5: \code{\link{stsp}}
  \item 6: \code{\link{pseries}}
  \item 7: \code{\link{zvalues}}
}

Laurent polynomials \code{\link{lpolm}} have a special place because they are detached from/unconnected to elements \code{\link{lmfd}}, \code{\link{rmfd}}, \code{\link{stsp}}, \code{\link{pseries}}, \code{\link{zvalues}}.

Operations with \code{\link{lpolm}} are only possible with \code{\link{polm}} objects (and those which can be coerced to \code{\link{polm}}).

Operations which are not defined are 

\itemize{
  \item \code{\link{pseries}} and \code{\link{zvalues}}: \code{\link{pseries}} cannot be coerced (easily) to \code{\link{zvalues}} object
  \item Anything involving \code{\link{lpolm}} except for \code{\link{polm}} objects
}
}
\examples{
p = test_polm(degree = 2)
lp = test_lpolm(degree_max = 1, degree_min = -1)
l = test_lmfd()
r = test_lmfd()
ss = test_stsp(s = 1)
ps = pseries(ss)
zv = zvalues(ss)

# Only one argument
rationalmatrices:::upgrade_objects(force = TRUE, l)
rationalmatrices:::upgrade_objects(force = FALSE, l)

# Upgrade poly to rmfd
rationalmatrices:::upgrade_objects(force = TRUE, p, r)
rationalmatrices:::upgrade_objects(force = TRUE, l, ps)
}
