% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zvalues.R
\name{zvalue}
\alias{zvalue}
\alias{zvalue.lpolm}
\alias{zvalue.polm}
\alias{zvalue.lmfd}
\alias{zvalue.rmfd}
\alias{zvalue.stsp}
\title{Frequency Response Function}
\usage{
zvalue(obj, z, ...)

\method{zvalue}{lpolm}(obj, z = NULL, ...)

\method{zvalue}{polm}(obj, z = NULL, ...)

\method{zvalue}{lmfd}(obj, z = NULL, ...)

\method{zvalue}{rmfd}(obj, z = NULL, ...)

\method{zvalue}{stsp}(obj, z = NULL, ...)
}
\arguments{
\item{obj}{(rational) matrix object, i.e. a \code{\link{polm}}, \code{\link{lpolm}}, \code{\link{lmfd}}, \code{\link{rmfd}}, \code{\link{stsp}} object 
or an object which may be coerced to a polynomial matrix with \code{polm(obj)}.
The default \code{S3} method first coerces the input argument \code{obj} to a \code{polm} object. 
If this fails an error is thrown.}

\item{z}{(numeric or complex) vector of length one at which to evaluate the rational matrix.}

\item{...}{optional additional parameters}
}
\value{
Matrix of dimensions of input object
}
\description{
This function evaluates a rational matrix at one given (complex) arguments. 
It is the non-plural version of \code{\link{zvalues}}.
}
