% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representation_zvalues.R
\name{zvalues}
\alias{zvalues}
\alias{zvalues.default}
\alias{zvalues.polm}
\alias{zvalues.lpolm}
\alias{zvalues.lmfd}
\alias{zvalues.rmfd}
\alias{zvalues.stsp}
\title{Frequency Response Function}
\usage{
zvalues(obj, z, f, n.f, sort.frequencies, ...)

\method{zvalues}{default}(obj, z = NULL, f = NULL, n.f = NULL, sort.frequencies = FALSE, ...)

\method{zvalues}{polm}(obj, z = NULL, f = NULL, n.f = 5, sort.frequencies = FALSE, ...)

\method{zvalues}{lpolm}(obj, z = NULL, f = NULL, n.f = 5, sort.frequencies = FALSE, ...)

\method{zvalues}{lmfd}(obj, z = NULL, f = NULL, n.f = 5, sort.frequencies = FALSE, ...)

\method{zvalues}{rmfd}(obj, z = NULL, f = NULL, n.f = 5, sort.frequencies = FALSE, ...)

\method{zvalues}{stsp}(obj, z = NULL, f = NULL, n.f = 5, sort.frequencies = FALSE, ...)
}
\arguments{
\item{obj}{(rational) matrix object, i.e. a \code{\link{polm}}, \code{\link{lpolm}}, \code{\link{lmfd}}, \code{\link{rmfd}}, \code{\link{stsp}} object. 
or an object which may be coerced to a polynomial matrix with \code{polm(obj)}.
The default \code{S3} method acts as a constructor for zvalues objects: and array object and \code{z} are supplied, corresponding to the evaluated object at the values of \code{z}. (Numeric or complex) vectors or matrices are coerced arrays. If \code{f} is supplied instead of \code{z}, then \code{z} is considered to be \code{exp(complex(imaginary = (-2*pi)*f))}.}

\item{z}{(numeric or complex) vector, points at which to evaluate the rational matrix.
Theses values are stored as attribute and can be accessed through \code{zvalues_obj$z}.}

\item{f}{(numeric) vector of frequencies. If \code{z = NULL} then \code{z} is set to 
\code{z = exp(complex(imaginary = (-2*pi)*f))}. If \code{z} is non \code{NULL}, then 
\code{f} is set to \code{f = -Arg(z)/(2*pi)}.
Theses values can be accessed through \code{zvalues_obj$f}.}

\item{n.f}{(integer) number of frequencies. If \code{z = f = NULL} then a grid of frequencies \code{f = (0:(n.f-1))/n.f} 
is used (and \code{z} is generated as explained above).}

\item{sort.frequencies}{boolean, sort frequencies}

\item{...}{optional additional parameters}
}
\value{
Object of type \code{zvalues}
}
\description{
This function evaluates a rational matrix at given (complex) arguments. 
If the rational matrix corresponds to a \emph{linear, dynamic filter}, then 
this rational function is also called \emph{transfer function} of the filter 
and the values of the function on the complex unit circle are called 
\emph{frequency response} of the filter.
}
\details{
A \code{zvalues} object is simply a (complex valued) 3-dimensional array of dimension \eqn{(m,n,n.f)} 
with attribute \code{z} and class attribute \code{c('zvalues','ratm')}. 
The dimensions \eqn{(m,n,n.f)} may also be zero, where \code{n.f} is the length of \code{z}.

\code{z} and \code{f} can be accessed through \code{zvalues_obj$z} and \code{zvalues_obj$z}, respectively.

If you only need the value of the matrix evaluated at one point, then use 
\code{\link{zvalue}(obj, z)}.
}
\examples{
(vv = zvalues(1:3, z = 4:6))
(mm = zvalues(matrix(1:4, 2, 2), z = 1))
(aa = zvalues(array(1:8, dim = c(2,2,2)), f = c(0.2, 0.8)))
(ll = test_lmfd(dim = c(2,2), bpoles = 1, bzeroes = 1))
zvalues(ll)
}
