% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representation_conversions.R
\name{as.lmfd}
\alias{as.lmfd}
\alias{as.lmfd.pseries}
\title{Coerce to Left Matrix Fraction Description}
\usage{
as.lmfd(obj, method, ...)

\method{as.lmfd}{pseries}(obj, method, ...)
}
\arguments{
\item{obj}{object}

\item{method}{character string}

\item{...}{optional additional arguments}
}
\value{
object of class \code{\link{lmfd}}
}
\description{
The function \code{\link{as.lmfd.pseries}} calls
\code{\link{pseries2lmfd}} with default parameters.
Of course the \code{\link{pseries}} object must contain
sufficiently many lags.
}
