% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_methods.R
\name{bind}
\alias{bind}
\alias{rbind.ratm}
\alias{cbind.ratm}
\title{Combine Rational Matrices by Rows or Columns}
\usage{
\method{rbind}{ratm}(...)

\method{cbind}{ratm}(...)
}
\arguments{
\item{...}{rational matrix objects, or objects which may be coerced to rational matrix objects.}
}
\value{
rational matrix object.
}
\description{
The function `rbind()` and `cbind()` take a sequence of rational matrix objects 
(i.e. \code{\link{polm}}, \code{\link{lmfd}}, \code{\link{stsp}}, \code{\link{pseries}} 
or \code{\link{zvalues}} objects) and combines these matrices by 
rows or columns.
}
\details{
The input matrices are first coerced to objects of the same class, 
as described for the group operator methods in \code{\link{Ops.ratm}}.
}
\examples{
a1 = test_polm(dim = c(2,3), degree = 2)
a2 = test_polm(dim = c(1,3), degree = 1)

rbind(a1, a2)                  # => polm object
rbind(lmfd(diag(2), a1), a2)   # => stsp object
rbind(a1, as.stsp(a2))         # => stsp object
rbind(a1, pseries(a2))         # pseries object 
rbind(zvalues(a1), a2)        # zvalues object

a2 = test_polm(dim = c(2,1), degree = 3)

cbind(a1, a2)
cbind(lmfd(diag(2), a1), a2)
cbind(a1, as.stsp(a2))
cbind(a1, pseries(a2))
cbind(zvalues(a1), a2)

# the following exmpales throw an error
\dontrun{
rbind(a1, a2)   # the number of columns does not coincide 
cbind(pseries(a1), zvalues(a2)) # there is no automatic coercion
                                 # from pseries to zvalues
}
}
