% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_methods.R
\name{extract}
\alias{extract}
\alias{[.polm}
\alias{[.lpolm}
\alias{[.lmfd}
\alias{[.rmfd}
\alias{[.stsp}
\alias{[.pseries}
\alias{[.zvalues}
\alias{$.lmfd}
\alias{$.rmfd}
\alias{$.stsp}
\alias{$.zvalues}
\title{Extract Parts of a Rational Matrix}
\usage{
\method{[}{polm}(x, i, j)

\method{[}{lpolm}(x, i, j)

\method{[}{lmfd}(x, i, j)

\method{[}{rmfd}(x, i, j)

\method{[}{stsp}(x, i, j)

\method{[}{pseries}(x, i, j)

\method{[}{zvalues}(x, i, j)

\method{$}{lmfd}(x, name)

\method{$}{rmfd}(x, name)

\method{$}{stsp}(x, name)

\method{$}{zvalues}(x, name)
}
\arguments{
\item{x}{a rational matrix, i.e. a \code{\link{polm}}, \code{\link{lpolm}}, 
\code{\link{lmfd}}, \code{\link{rmfd}}, \code{\link{stsp}}, 
\code{\link{pseries}} or \code{\link{zvalues}} object.}

\item{i, j}{indices (integer or boolean vector)}

\item{name}{character: A,B,C,D for \code{stsp} objects, 
a,b for \code{lmfd} objects, 
c,d for \code{rmfd} objects and 
z,f for \code{zvalues} objects.}
}
\value{
The subsetting operation \code{x[i,j]} returns a rational matrix of the same class 
as the input \code{x}. The mode of the output of the \code{$} operator depends on 
which "component" of the rational matrix is extracted. See the details above.
}
\description{
The subsetting operation \code{x[,]} for rational matrices works analogously to the subsetting of 
ordinary matrices. However, this operator is only partly implemented 
for \code{\link{lmfd}} and \code{\link{lmfd}} objects. See the details below.
\cr
The \code{$} operator may be used to extract the polynomial factors of a left/right 
matrix fraction description (\code{\link{lmfd}}, \code{\link{rmfd}} object). Furthermore  
one may retrieve the parameter matrices \eqn{A,B,C,D} of a state space representation 
(\code{\link{stsp}} object). For an \code{\link{zvalues}} object, we may 
access the complex numbers at which the rational matrix has been evaluated.
}
\details{
\itemize{
\item \code{x[]} or \code{x[,]} simply return the original object.
\item \code{x[i]} returns a "vector", i.e. an \eqn{(s,1)} dimensional matrix. 
\item \code{x[i,]}, \code{x[,j]} or \code{x[i,j]} return a rational matrix 
      with rows selected by \code{i} and columns selected by \code{j}. 
\item \strong{Note:} for \code{\link{lmfd}} objects (\code{\link{rmfd}} objects) 
      only extraction of columns (respectively rows) is implemented. Therefore, 
      e.g. \code{x[i,]} throws an error if \code{x} is an \code{lmfd} object.
\item Note that "named" arguments are not supported (in order to simplify the coding). 
\item In order to have a finer control, one may e.g. use \code{unclass(x)[,,]}.
\item \code{x$a}, \code{x$b} returns the left, respectively right factor of an 
      LMFD \eqn{x(z) = a^{-1}(z)b(z)}. (\code{x} is an \code{\link{lmfd}} object).
\item \code{x$c}, \code{x$d} returns the right, respectively left factor of an 
      RMFD  \eqn{x(z) = d(z)c^{-1}(z)}. (\code{x} is an \code{rmfd} object.)
\item \code{x$A}, \code{x$B}, \code{x$C}, \code{x$D} return the 
      parameter matrices of the statespace realization  
      \eqn{x(z) = D + z (I- Az)^{-1}B}. (\code{x} is an \code{stsp} object.)
\item If \code{x} is an \code{\link{zvalues}} object, then 
      \code{x$z} returns the vector of complex points at which the 
      rational matrix \eqn{x} has been evaluated. Furthermore \code{x$f} gives 
      the corresponding "frequencies", i.e. \code{x$f = -Arg(x$z)/(2*pi)}.
}
}
\examples{
# polynomial matrices 
a = test_polm(dim = c(3,2), degree = 1)
a[]          # returns a
a[,]         # returns a
a[c(1,3,6)]  # returns a "vector" with the (1,1), (3,1) and (3,2) element of a
a[1,]        # returns the first row of a 
a[,2]        # returns the second column of a 
a[c(TRUE,FALSE,TRUE),c(FALSE, TRUE)] # returns a 2 by 1 matrix 
a[c(1,1),c(2,1)] # returns a 2 by 2 matrix 
# check with pseries/zvalues
all.equal(pseries(a[c(1,1),c(2,1)]), pseries(a)[c(1,1),c(2,1)])
all.equal(zvalues(a[c(1,1),c(2,1)]), zvalues(a)[c(1,1),c(2,1)])

\dontrun{
a[i=1, j=2] # throws an error, since "named" arguments are not allowed.
}

# the subsetting operator [,] is only implemented for "lmfd" columns
(l = test_lmfd(dim = c(2,2), degrees = c(1,1)))
l[,1]

# the subsetting operator [,] is only implemented for "rmfd" rows
(r = test_rmfd(dim = c(2,2), degrees = c(1,1)))
r[1,]
}
