% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_kronecker.R
\name{Kronecker-Indices}
\alias{Kronecker-Indices}
\alias{basis2nu}
\alias{nu2basis}
\alias{pseries2nu}
\title{Tools related to Kronecker indices}
\usage{
basis2nu(basis, m)

nu2basis(nu)

pseries2nu(obj, Hsize = NULL, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{basis}{s-dimensional (integer) vector which contains the indices of the
basis rows of the Hankel matrix.}

\item{m}{(integer) the number of rows of the underlying rational matrix.}

\item{nu}{vector with the Kronecker indices, i.e. an m-dimensional integer vector.}

\item{obj}{\code{\link{pseries}} object or 3-D array with dimension \eqn{(m,n,l+1)}. 
This object represents the impulse response of a rational matrix.}

\item{Hsize}{integer vector \code{c(f,p)}, number of block rows and block columns 
of the Hankel matrix of the impulse response coefficients. 
      If NULL a default choice is made.}

\item{tol}{tolerance parameter, used by \code{\link{qr}}.}
}
\value{
\item{basis2nu}{returns the Kronecker indices (\code{nu}) for given indices of the basis rows.}
\item{nu2basis}{returns the indices of the basis rows for given Kronecker indices \code{nu}.}
\item{pseries2nu}{determines the Kronecker indices for given impulse response  coefficients.
The function uses a QR decomposition (with pivoting) of the transpose of 
the Hankel matrix to determine the rank and the basis for its row space. See \code{\link{qr}}.}
}
\description{
The Kronecker indices \eqn{(\nu_1,\ldots,\nu_m)}{(nu[1],\ldots,nu[m])} describe
a (nice) basis of the row space of the Hankel matrix of the impulse response
coefficients. These indices are e.g. used to construct a (unique) LMFD representation
in echelon canonical form for a given impulse response.
}
\details{
The function \code{pseries2nu} first constructs a Hankel matrix of the impulse response coefficients 
with \eqn{f} and \eqn{p} block rows and columns respectively. Then a (nice) basis for the row space is 
computed via a QR decomposition (with pivoting) of the transposed Hankel matrix. See \code{\link{qr}}.
If the size of the Hankel matrix is not specified (the parameter \code{Hsize=c(f,p)} is missing),  
then a default choice is made such that \eqn{f+p-1 = l}, \eqn{p\geq 1}{p \ge 1} and 
\eqn{f \geq p+1}{f \ge p+1} holds.
\cr
The function \code{pseries2nu} throws an error if the conditions \eqn{p\geq 1}{p \ge 1},  
\eqn{f \geq 2}{f \ge 2} and \eqn{l \geq f+p-1}{l \ge f+p-1} 
are not satisfied. In particular, this implies that 
\eqn{l \geq 2}{l \ge 2} must hold.
}
\examples{
basis = c(1,2,3,4,6,7)     # suppose rows 1,2,3,4,6,7 of the Hankel matrix form a basis
nu = basis2nu(basis, m=3)  # compute the corresponding Kronecker index
print(nu)
all.equal(basis,nu2basis(nu))  # nu2basis(nu) returns the indices of the basis rows (basis)

# generate random rational matrix in statspace form
# make sure that the matrix is stable
m = 3
n = 2
s = 7
A = matrix(rnorm(s*s), nrow = s, ncol = s)
A = A / (1.1 * max(abs(eigen(A, only.values = TRUE)$values)))
x = stsp(A, B = matrix(rnorm(s*n), nrow = s, ncol = n),
            C = matrix(rnorm(s*m), nrow = m, ncol = s),
            D = diag(1, nrow = m, ncol = n))
k = pseries(x, lag.max = 20)

# compute the Kronecker indices of this  rational matrix
pseries2nu(k)

\dontrun{
# Suppose the rational matrix has dimension m=2. Then the rows 1,2,5
# do not form a "nice" basis for the row space of the Hankel matrix.
# Therefore "basis2nu" stops with an error message.
basis2nu(c(1,2,5), m=2)
}

}
\references{
\insertRef{Hannan.Deistler12}{rationalmatrices}
}
