% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utilities_print.R
\name{print methods}
\alias{print methods}
\alias{print.lpolm}
\alias{print.polm}
\alias{print.lmfd}
\alias{print.rmfd}
\alias{print.stsp}
\alias{print.pseries}
\alias{print.zvalues}
\title{Print Methods}
\usage{
\method{print}{lpolm}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{polm}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{lmfd}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{rmfd}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{stsp}(x, digits = NULL, ...)

\method{print}{pseries}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{zvalues}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)
}
\arguments{
\item{x}{rational matrix object, i.e. a \code{\link{polm}}, \code{\link{lpolm}}, \code{\link{lmfd}}, \code{\link{rmfd}},
\code{\link{stsp}}, \code{\link{pseries}} or \code{\link{zvalues}} object.}

\item{digits}{(integer) if non \code{NULL} then correspondingly rounded numbers are printed, 
see \code{\link{round}}.}

\item{format}{(character string) selects specific output formats. Note that 
\code{\link{stsp}} objects have no format option. The option \code{'character'} 
is only implemented for polynomials, Laurent polynomials, LMFDs and RMFDs with real coefficients, se}

\item{...}{Further parameters are ignored.}
}
\value{
\code{invisible(x)}
}
\description{
Printing rational matrix objects.
}
\examples{
# for polynomials six different print formats are implemented ###################
a = test_polm(dim = c(2,3), degree = 2, random = TRUE)

for (fmt in c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character")) {
   cat('\nformat =', fmt, '\n')
   print(a, digits = 2, format = fmt)
}

# "empty" (2 x 0) polynomial matrix (degree = 2)
a = test_polm(dim = c(2,0), degree = 0)
print(a)

# random (2 x 1) polynomial matrix with complex coefficients (degree = 2)
a = polm(array(complex(real = stats::rnorm(2*1*3), 
                       imaginary = stats::rnorm(2*1*3)), dim = c(2,1,3)))
print(a, digits = 2)
\dontrun{
# the format option 'character' is only implemented for polynomials matrices 
# with real coefficients!
print(a, digits = 2, format = 'character')
}

# print a rational matrix in statespace form
a = test_stsp(dim = c(3,3), s = 2)
print(a, digits = 2)

# print a rational matrix in 'lmfd' form 
a = test_lmfd(dim = c(2,3), degrees = c(2,1))
print(a, digits = 2, format = 'character')

# print impulse response 
print(pseries(a), format = 'i|zj', digits = 2)

# print frequency response 
print(zvalues(a), format = 'iz|j', digits = 2)

}
