% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_kronecker.R
\name{pseries2hankel}
\alias{pseries2hankel}
\title{Construct Hankel Matrix from Impulse Response Coefficients}
\usage{
pseries2hankel(obj, Hsize = NULL)
}
\arguments{
\item{obj}{\code{\link{pseries}} object or 3-D array with dimension \eqn{(m,n,l+1)}.}

\item{Hsize}{integer vector \code{c(f,p)}, number of block rows and block columns. 
If NULL a default choice is made.}
}
\value{
Block Hankel matrix with dimension \eqn{(fm, pn)} and 
        attributes \code{order=c(m,n,f,p)} and \code{k0}. The \eqn{(m,n)}-dimensional 
        matrix \code{k0} is the lag zero coefficient of the impulse response.
}
\description{
This (internal) helper function builds a Hankel matrix from a 
given impulse response of dimension \eqn{(m,n)} with 
\eqn{l} lags. If the parameter \code{Hsize=c(f,p)} is not given, 
then a default choice for the number of block rows (\eqn{f}) and 
block columns (\eqn{p}) is made such that 
\eqn{f+p-1 = l}, \eqn{p\geq 1}{p \ge 1} and 
\eqn{f \geq p+1}{f \ge p+1}.
\cr
The function throws an error if the conditions \eqn{p\geq 1}{p \ge 1},  
\eqn{f \geq 2}{f \ge 2} and \eqn{l \geq f+p-1}{l \ge f+p-1} 
are not satisfied. In particular, this implies that 
\eqn{l \geq 2}{l \ge 2} must hold.
}
\keyword{internal}
