% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rmfd}
\alias{rmfd}
\title{Constructor for Right Matrix Fraction Descriptions (RMFDs)}
\usage{
rmfd(c = NULL, d = NULL)
}
\arguments{
\item{c, d}{\code{\link{polm}} objects, or objects which may be coerced to a \code{\link{polm}} object, via \code{x = polm(x)}. 
Either of the two arguments may be omitted.}
}
\value{
An object of class \code{rmfd}.
}
\description{
A Right Matrix Fraction Description (RMFD) of a rational matrix, \eqn{k(z)}
say, is  pair \eqn{(c(z),d(z))} of polynomial matrices, such that
\eqn{k(z) = d(z) c^{-1}(z)}. The polynomial matrix \eqn{c(z)} must be square
and invertible.
}
\details{
Suppose that \eqn{k(z) = d(z) c^{-1}(z)} is an \eqn{(m,n)}-dimensional matrix
and that \eqn{c(z)} and \eqn{d(z)} have degrees \eqn{p} and \eqn{q}
respectively. The corresponding \code{rmfd} object stores the coefficients of
the polynomials \eqn{c(z), d(z)} in an \eqn{(n(p+1)+m(q+1), n)} dimensional
(real or complex valued) matrix. The \code{rmfd} object also stores the
attribute \code{order = c(m,n,p,q)} and a class attribute \code{c("lmfd",
"ratm")}.

For a valid RMFD we require \eqn{m>0} and \eqn{p\geq 0}{p \ge 0}.
}
\examples{
### (1 x 1) rational matrix k(z) =  (3+2z+z^2) (1+z+z^2)^(-1)
rmfd(c(1,1,1), c(3,2,1)) \%>\% print(format = 'c')

### (1 x 1) rational matrix k(z) = (3+2z+z^2)
rmfd(c = c(3,2,1)) \%>\% print(format = 'c')

### (1 x 1) rational matrix k(z) = (1+z+z^2)^(-1)
rmfd(c(1,1,1)) \%>\% print(format = 'c')

### (3 x 2) rational matrix with degrees p=1, q=1
x = rmfd(array(rnorm(2*2*2), dim = c(2,2,2)), 
         array(rnorm(3*2*2), dim = c(3,2,2)))
is.rmfd(x)
dim(x)
str(x)
print(x, digits = 2)

xr = test_rmfd(dim = c(3,2), degree = c(2,2))

\dontrun{
### the following calls to rmfd() throw an error 
rmfd() # no arguments!
rmfd(c = test_polm(dim = c(2,3), degree = 1))  # c(z) must be square 
rmfd(c = test_polm(dim = c(2,2), degree = -1)) # c(z) must have degree >= 0
}
}
\seealso{
Useful methods and functions for the \code{rmfd} class are: 
\itemize{
   \item \code{\link{test_rmfd}} generates random rational matrices in RMFD form.
   \item checks: \code{\link{is.rmfd}}, \code{\link{is.miniphase}}, \code{\link{is.stable}} 
         and \code{\link{is.coprime}}.
   \item generic S3 methods: \code{\link[rationalmatrices]{dim}}, 
         \code{\link[rationalmatrices]{str}} and \code{\link[rationalmatrices]{print}}.
   \item arithmetics: \code{\link[rationalmatrices]{Ops.ratm}}.
   \item matrix operations: \code{\link[rationalmatrices]{bind}}.
   \item extract the factors \eqn{c(z)} and \eqn{d(z)} with \code{\link{$.rmfd}}.      
   \item \code{\link{zeroes}}, \code{\link{pseries}}, \code{\link{zvalues}}, ...  
}
}
